+++ adminforth/spa/src/views/ListView.vue
@@ -110,7 +110,6 @@
 import BreadcrumbsWithButtons from '@/components/BreadcrumbsWithButtons.vue';
 import ResourceListTable from '@/components/ResourceListTable.vue';
 import { useCoreStore } from '@/stores/core';
-import { useModalStore } from '@/stores/modal';
 import { useFiltersStore } from '@/stores/filters';
 import { callAdminForthApi, getIcon } from '@/utils';
 import { initFlowbite } from 'flowbite';
@@ -180,6 +179,16 @@ async function getList() {
 }
 
 async function startBulkAction(actionId) {
+  const action = coreStore.resource.options.bulkActions.find(a => a.id === actionId);
+  if (action.confirm) {
+    const confirmed = await window.adminforth.confirm({
+      message: action.confirm,
+    });
+    if (!confirmed) {
+      return;
+    }
+  }
+
   const data = await callAdminForthApi({
     path: '/start_bulk_action',
     method: 'POST',
@@ -193,6 +202,14 @@ async function startBulkAction(actionId) {
   if (data?.ok) {
     checkboxes.value = [];
     await getList();
+
+    if (data.successMessage) {
+      window.adminforth.alert({
+        message: data.successMessage,
+        variant: 'success'
+      });
+    }
+
   }
   if (data?.error) {
     showErrorTost(data.error);