+++ adminforth/spa/src/views/LoginView.vue
@@ -56,6 +56,19 @@
                             </button>
                         </div>
 
+                        <div class="flex items-start mb-5 hidden">
+                          <div class="flex items-center h-5">
+                            <input id="remember" type="checkbox" value="" class="w-4 h-4 border border-gray-300 rounded bg-gray-50 focus:ring-3 focus:ring-blue-300 dark:bg-gray-700 dark:border-gray-600 dark:focus:ring-blue-600 dark:ring-offset-gray-800 dark:focus:ring-offset-gray-800" required />
+                          </div>
+                          <label for="remember" class="ms-2 text-sm font-medium text-gray-900 dark:text-gray-300">Remember me</label>
+                        </div>
+                        
+                        <component 
+                          v-for="c in coreStore?.config?.loginPageInjections?.underInputs || []"
+                          :is="getCustomComponent(c)"
+                          :meta="c.meta"
+                        />
+
                         <div v-if="error" class="flex items-center p-4 mb-4 text-sm text-red-800 rounded-lg bg-red-50 dark:bg-gray-800 dark:text-red-400" role="alert">
                           <svg class="flex-shrink-0 inline w-4 h-4 me-3" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 20 20">
                             <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5ZM9.5 4a1.5 1.5 0 1 1 0 3 1.5 1.5 0 0 1 0-3ZM12 15H8a1 1 0 0 1 0-2h1v-3H8a1 1 0 0 1 0-2h2a1 1 0 0 1 1 1v4h1a1 1 0 0 1 0 2Z"/>
@@ -65,6 +78,7 @@
                           {{ error }}
                           </div>
                         </div>
+                        
                         
 
                         <div v-if="coreStore.config?.loginPromptHTML"
@@ -101,6 +115,7 @@
 
 <script setup>
 
+import { getCustomComponent } from '@/utils';
 import { onMounted, ref, computed } from 'vue';
 import { useCoreStore } from '@/stores/core';
 import { useUserStore } from '@/stores/user';