+++ Changelog.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.22] - 2024-09-25
+
+### Improved
+
+- close all toasts automatically if during route change they exist more then 5 seconds. 
+- Also for alerts without timeout param set 30 seconds as default timeout, you still can use 'unlimited' intentially to never hide alert 
+- Add alert when record is created
+
+### Security fixes
+
+- When user was nto found during login, return same message as when password is wrong to prevent email/username enumeration
+
+
+## [1.3.21] - 2024-09-25
+
+### Added
+
+- Added underInputs injection to login page
+- now plugins can issue a JWT tokens without users (to not expose user id in the token), userless verification achieved by setting decodeUser param in auth.verify to false
+
 ## [1.3.18] - 2024-09-20
 
 ### Fixed